New component MainContent
