+++ src/components/main-content/MainContent.module.css
@@ -1,0 +1,4 @@
+.contentWrapper {
+    width: 1080px;
+    margin: 0 auto;
+}
